Move `harp::test` to `harp::fixtures`
