+++ crates/harp/src/environment.rs
@@ -281,7 +281,7 @@ mod tests {
     use crate::exec::RFunctionExt;
     use crate::object::r_length;
     use crate::r_symbol;
-    use crate::test::r_task;
+    use crate::fixtures::r_task;
 
     fn new_test_environment(hash: bool) -> Environment {
         let test_env = RFunction::new("base", "new.env")
