+++ crates/harp/src/lib.rs
@@ -13,6 +13,7 @@ pub mod error;
 pub mod eval;
 pub mod exec;
 pub mod external_ptr;
+pub mod fixtures;
 pub mod format;
 pub mod json;
 pub mod library;
@@ -33,7 +34,6 @@ pub mod string;
 pub mod symbol;
 pub mod sys;
 pub mod table;
-pub mod test;
 pub mod traits;
 pub mod utils;
 pub mod vec_format;
@@ -58,12 +58,12 @@ pub use harp::exec::top_level_exec;
 pub use harp::exec::try_catch;
 pub use harp::exec::try_eval;
 pub use harp::exec::try_eval_silent;
+pub use harp::fixtures::r_task;
+pub use harp::fixtures::IS_TESTING;
 pub use harp::object::list_get;
 pub use harp::object::list_poke;
 pub use harp::object::RObject;
 pub use harp::symbol::RSymbol;
-pub use harp::test::r_task;
-pub use harp::test::IS_TESTING;
 pub use harp::utils::get_option;
 pub use harp_macros::register;
 
@@ -251,9 +251,9 @@ mod tests {
     use libr::*;
 
     use super::*;
+    use crate::fixtures::r_task;
     use crate::object::RObject;
     use crate::protect::RProtect;
-    use crate::test::r_task;
     use crate::utils::r_is_null;
     use crate::utils::r_typeof;
